requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
segmentation-models-pytorch>=0.3.0

# Geospatial
rasterio>=1.3.0
geopandas>=0.13.0
shapely>=2.0.0
pyproj>=3.6.0
geemap>=0.25.0
earthengine-api>=0.1.350

# Data Science
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Web App
streamlit>=1.28.0
streamlit-folium>=0.15.0
plotly>=5.17.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
albumentations>=1.3.0
scipy>=1.11.0

